Fix Double/CGFloat arithmetic mismatches in three Canvas-based boards
The fileprivate fix cleared completely -- no access-level or availability
errors anywhere. Two new, real errors surfaced, both in
TVDuelGameBoards.swift:

  TVDuelGameBoards.swift:552:67: error: ambiguous use of operator '-'
  TVDuelGameBoards.swift:422:25: error: the compiler is unable to
    type-check this expression in reasonable time

The first is a genuine type bug: vm.state.strikerX is Double (it decodes
from server JSON), scale is CGFloat, and Swift has no automatic
conversion between the two distinct concrete types -- only literals
coerce automatically. Auditing every `* scale` (and the equivalent `*
cell`/`* tile` multipliers elsewhere) turned up the identical mismatch in
several more places the compiler hadn't reported yet in this same pass:
AirHockey's paddle/puck math and two `.frame()` calls, and Carrom's coin
positions and both board-size frames. Brick Breaker
(TVSoloGameBoards.swift) had the same pattern for its bricks, ball, and
paddle. Wrapped every Double-typed operand in CGFloat(...) at the point
of multiplication -- Neon Snake's `cell` usage and 2048's `tile` usage
were already written correctly and needed no change.

The second error (Ludo's body "unable to type-check in reasonable time")
is left unaddressed for now: it's defined in this same file, compiled in
the same seven-file batch as the Carrom ambiguity, and ambiguous
overloads are a known way to cascade into type-checker timeouts
elsewhere in the same unit. Pushing the definite fix first to see
whether that was the actual cause before spending a separate fix on
Ludo's view complexity.

+++ ios/GameLabTV/Views/Games/TVDuelGameBoards.swift
@@ -285,24 +285,26 @@ struct TVAirHockeyBoardView: View {
                                                       width: 120, height: 120)),
                                with: .color(.white.opacity(0.15)), lineWidth: 2)
 
-                    // Paddles
+                    // Paddles. state fields decode as Double (server JSON), scale is
+                    // CGFloat -- Swift has no automatic Double<->CGFloat conversion,
+                    // so each Double sub-expression is wrapped once before scaling.
                     for (i, p) in vm.state.paddles.enumerated() {
-                        let y: CGFloat = i == 0 ? 6 * scale : (vm.state.height - 6) * scale
+                        let y: CGFloat = i == 0 ? 6 * scale : CGFloat(vm.state.height - 6) * scale
                         ctx.fill(
                             Path(roundedRect: CGRect(
-                                x: (p.x - vm.state.paddleWidth / 2) * scale, y: y - 8,
-                                width: vm.state.paddleWidth * scale, height: 16),
+                                x: CGFloat(p.x - vm.state.paddleWidth / 2) * scale, y: y - 8,
+                                width: CGFloat(vm.state.paddleWidth) * scale, height: 16),
                                  cornerRadius: 8),
                             with: .color(i == 0 ? .cyan : .pink))
                     }
 
                     ctx.fill(
-                        Path(ellipseIn: CGRect(x: (vm.state.puck.x - 3) * scale,
-                                               y: (vm.state.puck.y - 3) * scale,
+                        Path(ellipseIn: CGRect(x: CGFloat(vm.state.puck.x - 3) * scale,
+                                               y: CGFloat(vm.state.puck.y - 3) * scale,
                                                width: 6 * scale, height: 6 * scale)),
                         with: .color(.white))
                 }
-                .frame(width: vm.state.width * scale, height: vm.state.height * scale)
+                .frame(width: CGFloat(vm.state.width) * scale, height: CGFloat(vm.state.height) * scale)
 
                 if vm.state.finished {
                     Text("GAME OVER").font(.system(size: 54, weight: .heavy)).tracking(5)
@@ -311,7 +313,7 @@ struct TVAirHockeyBoardView: View {
                         .background(RoundedRectangle(cornerRadius: 20).fill(.black.opacity(0.8)))
                 }
             }
-            .frame(width: vm.state.width * scale + 8, height: vm.state.height * scale + 8)
+            .frame(width: CGFloat(vm.state.width) * scale + 8, height: CGFloat(vm.state.height) * scale + 8)
         }
         .onAppear { vm.bind(roomCode: room.code) }
     }
@@ -520,7 +522,7 @@ struct TVCarromBoardView: View {
             ZStack {
                 RoundedRectangle(cornerRadius: 14)
                     .fill(Color(hex: "d9b382"))
-                    .frame(width: vm.state.board * scale, height: vm.state.board * scale)
+                    .frame(width: CGFloat(vm.state.board) * scale, height: CGFloat(vm.state.board) * scale)
                     .overlay(RoundedRectangle(cornerRadius: 14)
                         .stroke(Color(hex: "6b4f2a"), lineWidth: 10))
 
@@ -538,22 +540,24 @@ struct TVCarromBoardView: View {
                                                       width: 90, height: 90)),
                                with: .color(Color(hex: "6b4f2a").opacity(0.5)), lineWidth: 3)
 
+                    // coin.x/y and strikerX decode as Double (server JSON); scale is
+                    // CGFloat, so each is wrapped before scaling -- see AirHockey above.
                     for coin in vm.state.coins {
                         let color: Color = coin.kind == "queen" ? .red
                                          : coin.kind == "black" ? .black
                                          : Color(hex: "f5e6c8")
-                        ctx.fill(Path(ellipseIn: CGRect(x: coin.x * scale - 12,
-                                                        y: coin.y * scale - 12,
+                        ctx.fill(Path(ellipseIn: CGRect(x: CGFloat(coin.x) * scale - 12,
+                                                        y: CGFloat(coin.y) * scale - 12,
                                                         width: 24, height: 24)),
                                  with: .color(color))
                     }
                     // Striker
-                    ctx.fill(Path(ellipseIn: CGRect(x: vm.state.strikerX * scale - 16,
+                    ctx.fill(Path(ellipseIn: CGRect(x: CGFloat(vm.state.strikerX) * scale - 16,
                                                     y: 92 * scale - 16,
                                                     width: 32, height: 32)),
                              with: .color(.cyan))
                 }
-                .frame(width: vm.state.board * scale, height: vm.state.board * scale)
+                .frame(width: CGFloat(vm.state.board) * scale, height: CGFloat(vm.state.board) * scale)
             }
             Spacer()
             TVScoreStrip(players: vm.state.players)
